* Make a dedicated mapping for production mode * Run real smoke tests only for Chrome.
